#16 damage connected, bugfixes

+++ domain/src/main/java/ru/frozenpriest/pokebase/domain/pokemon/AddPokemonUseCase.kt
@@ -17,7 +17,7 @@ class AddPokemonUseCaseImpl @Inject constructor(
             PokemonDataRequest(
                 pokemonData.name,
                 pokemonData.level,
-                pokemonData.species.speciesId
+                pokemonData.species.speciesId.toInt()
             )
         )
     }
